pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-weather"
version = "1.0.0"
description = "MCP server for Taiwan weekly weather forecast (CWA Open Data)"
authors = [{name = "BrightSu"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.13.1",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]

[project.scripts]
mcp-weather = "server:main"
mcp-weather-http = "http_server:run_http_server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
